Fix build with Xcode 3.2.6.
Xcode 3.2.6 comes with a MacOS X 10.6 SDK that doesn't define the kAudioFormatMPEG4AAC_ELD enum value while newer versions of the SDK shipped with Xcode 4.x define it.

To avoid breakage when building with Xcode 3.2.6 (which is the last version that supported PPC builds), rename the fallback definitions so they work regardless of the SDK defining the enum values or not.

See issue #12 and PR #13 on GitHub.